Requirement: #4133 - Rework Duplicate Sop Policies
Interim commit (Not sure why all generated model files are marked as "modified" almost all of them aren't. Check them in anyway)

Modified implementation in previous commit. Changed the DuplicateSopHandler used by the dicom service to save the duplicate in the "Reconcile" folder instead. It also inserts a study process work queue entry for the duplicate instead. The study process processor will compare the duplicates against the existing sops and insert an SIQ entry if they are different.

- Added timestamp to the association to allow grouping of received instances by associations.
- Added GroupID to the WorkQueueUid. For duplicates, this field is tied to the association and is used as the duplicate storage folder name. RelativePath of the WorkQueueUid will be the relative path from the storage folder to the duplicate file.
- Modified DirectoryUtility.DeleteIfEmpty() to delete empty folders.

+++ ImageServer/Common/CommandProcessor/DeleteFileCommand.cs
@@ -107,12 +107,14 @@ namespace ClearCanvas.ImageServer.Common.CommandProcessor
             {
                 if (_restored)
                 {
-                    File.Delete(_backupFile);
+                    if (File.Exists(_backupFile))
+                        File.Delete(_backupFile);
                 }
             }
             else
             {
-                File.Delete(_backupFile);
+                if (File.Exists(_backupFile))
+                    File.Delete(_backupFile);
             }
         }
